requirements.txt
Django==5.0.4
djangorestframework==3.15.2
django-cors-headers==4.2.0
django-allauth==65.3.1
django-bootstrap4==24.4
psycopg2==2.9.10
gunicorn==22.0.0
numpy==2.2.4
pandas==2.2.3
scikit-learn==1.3.2
matplotlib==3.10.1
seaborn==0.13.2
pillow==11.0.0
xgboost==3.0.0
redis==5.2.1
django-dramatiq==0.13.0
imbalanced-learn==0.11.0
boto3==1.34.69
pytest==7.4.3
pytest-cov==4.1.0
django-s3-storage==1.14.6
daft==0.5.10
smart_open==6.2.0